factor runner into separate (testable) file
